LayoutTests/fast/forms/time/time-editable-components/time-editable-components-mouse-events-expected.txt

Test for mouse events for <input type=time>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Enabled Input

PASS input.value is "09:30"
PASS input.value is "09:12"
PASS input.value is "21:12"
PASS input.value is "18:12"
PASS input.value is "18:12"
PASS clickEventsFired is 4

Disabled Input

PASS clickEventsFired is 0

Readonly Input

PASS clickEventsFired is 4
PASS successfullyParsed is true

TEST COMPLETE
